requirements.txt
aiohttp==3.10
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.3.2
Flask-Cors==3.0.10
frozenlist
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
gunicorn==20.1.0
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict
openai==1.68.2
python-dotenv==1.0.0
requests==2.29.0
six==1.16.0
tqdm==4.65.0
urllib3==1.26.15
Werkzeug==2.3.3
yarl==1.9.2
zipp==3.15.0
litellm
